Add an overview guide for using the web based GUI to the Operations Guide.
With the introduction of a CLI tool for Brooklyn, the Getting Started Guide will
be updated to document using the CLI instead of the GUI.  This change is to copy
the content related to using the existing web based GUI from the Getting Started
Guide to the Operations Guide.  Incorporates some re-wording and changing of links
to reflect the new location.